pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "nyc_taxi"
version = "0.1.1"
authors = [
    {name = "Ali Shaheed", email = "contact@ali-fahad-shaheed.uk"},
]
description = "NYC Taxi data project"
dependencies = ["requests", "pandas", "python-dotenv",]

[tool.setuptools.packages]
find = {}
